refactor(ui-primitives): convert Badge, AlertDialog, Sheet to JSX [#812] [#813] [#814]
Replace imperative document.createElement with JSX element creation.
Use JSX event props (onClick, onKeydown, onPointerdown/up) instead of
addEventListener. Badge is now pure JSX. AlertDialog and Sheet use JSX
for element creation with imperative state management for
animation-aware visibility and focus trapping.
